update KEBC EB period histogram
Now matches result claimed by Armstrong et al 2014: that to 2σ
confidence (and marginalizing over all CBP periods), the occurrence
rate of coplanar CBPs about 5-10day period EBs is less than that about
>10day period EBs.

Unsurprisingly, when you cut out more short-period targets in the
histogram, the median moves to higher periods.

Note (again) though that this histogram is a not-at-all-precise way to
be talking about the occurrence rates of the planets -- the important
thing is how the "debiasing" is done, and their methodology does seem
reasonable.