add find_dips and dipsearchplot
We now have a (barely) functioning high-level module for the "recovery"
in injection-recovery: `find_dips`. It runs after all the detrending on
the SAP/PDC lightcurves.

Currently find_dips only runs BLS. On the TODO is adding routines that
do not assume periodicity of transits -- key if this code is to be
extended to more detached binaries.

Also on the TODO is to make dipsearchplot look better (and fold at the
appropriate periods, a current bug), and eventually to model-compare and
minimize "me-looking-at-lightcurves" time.