inj_recov: add P,t_0 finding from blsresult.
The previous approach of fitting a spline to phase-binned timeseries
during the plotting routines was A) not properly saving data, B) taking
too long, C) getting inaccurate epochs.

BLS finds epochs (in phase). Nominally, they should be stored in the
results that we get back from Kovacs' eebls routine. The issue is that
when parallelized, each workeronly returns the epoch from the
best-fitting period in its assigned frequency chunk.

To workaround this (rather than learn fortran and do something smart
like modify eebls.f to return better-select arrays and statistics), I
gave the `dipfind` dict in `inj_recov.find_dips` "coarsebls" and
"finebls" keys. coarsebls has blsresults from each worker's frequency
chunk. To get the better epochs, "finebls" runs narrow serial bls
windows over the nbestperiods to get the BIN INDEX at ingress of the
best transit (over that window) to that of egress.

This is a somewhat janky solution, but it seems to be working.