add realistic injected signal parameters
Drawing mostly from Foreman-Mackey et al. (2015)'s distributions, but
to skirt the issue of the stellar mass and radius, which are needed to
get a/R_star, I made the simple assumption Mtot=2Msun, Rstar=1.5Rsun.

This is reasonable-ish, but it also means "realistic" (single-star
planet) planet radius distributions push the typical transit depth to a
median of 0.1%. In other words, my EB subtraction will likely need to
improve.