add nb for possible EB subtraction methods
Includes tests of the Butterworth filter (from signal processing
literature), and of iterative whitening. The results are stored in
/results/eb_subtraction_diagnostics/butter_demo and /iterwhiten_demo
respectively.

An idea here (not pursued) was constructing Fourier filters from the
periodogram results. Not pursued A) because boxcar train's FT is sinc,
which has too-wide wings, and you don't know a-priori which peaks are
EB vs planet signals, and B) because Fourier methods rely on even
sampling, which even Kepler ~85% throughput aint.