add iterative whitening: PDM->fit->diff->repeat
We have an autocorrelated timeseries with both periodic and aperiodic
components. Dealing with the aperiodic components _may not be needed_
for the main result -- completeness down to ~<4Re CBPs in interesting
orbital period ranges -- to hold.

Going this direction is betting that removing the periodic components of
the EB will be a sufficient way to drive everything below a RMS of say
0.05% (500ppm). This is necessary to reach the aforementioned level of
completeness.

While aperiodic variability (e.g., attributed to starspots or magnetic
activity) could be important, fitting these kinds of models are trickier
(see: ARMA, ARIMA, ARFIMA, and GP regression). Ideally, it will not be
necessary.