add ability to search for real dips
I seriously attempted to maintain backwards compatibility, but haven't
tested whether or not that worked yet. So currently

>>> python run_the_machine.py -frd

works. But

>>> python run_the_machine.py -ir -N 100

will likely return bugs. This commit includes:

* update dipsearchplot for positive depths only, and no injections

* update write_search_result to deal with realsearch (including positive
depths only)

* update run_the_machine to have a recover only routine
(run_the_machine.recov). This was needed because the control flow
logic was too complicated in run_the_machine.injrecov to give it
dual-functionality.

* update todo and done lists