add summarize_realsearch_result
Sorts by something close to phase-folded SNR. I say "something close to"
because it's really using recovered transit depths (not true transit
depths), a noise which is just the RMS of the flux series, and the
number of transits which is just the observing baseline divided by the
period of the proposed CBP. So it's quite coarse.

However, plots show it does pretty well as an identifier of "what stands
a chance of ever being detected". So, sorting by phase-folded SNR, and
cutting e.g., at SNR>3 makes a reasonable list of objects (for now).

Also:

* bugfix control flow when whitened 6row plot err