save disk space in inj/recov by writing fewer pkls
At 30Mb/ea, ~1600 LCs means main run produces 48Gb of data.  That's
fine. (Even at ~60Mb/ea, accounting for allq, 100Gb is ok).

For injection recovery, we need far less data for production level runs.
We just need the planet and star params and whether the injected dip was
found. So we can do ~20k inj/recovs, and keep maybe the first ~50
pickles or something like that. (As sanity checks). We could also only
produce plots for those 50. After that, proceed by just keeping track of
some waaaay lighter-weight thing.

This is OFC risky b/c if there's a mistake in the information that is
saved (like "oh hey, I actually wanted all those inj/recov
periodograms!") then it all needs to be recomputed. However for the
number of inj/recovs we're talking about, we just cannot store the full
periodograms, and similarly there is not much point in holding onto the
full LC.

This is obviously different for the realsearch case, where we will keep
all of these intermediate data products.